+++ R/dummy_data.R
@@ -266,6 +266,7 @@ dummy_data <- function(no_obs    = 25000,
 
     # Extract NUTS2 region from NUTS3 code
     dummy_temp[["NUTS2"]] <- as.integer(substr(dummy_temp[["NUTS3"]], 3, 3))
+    dummy_temp[["NUTS3"]] <- as.integer(dummy_temp[["NUTS3"]])
 
     if (insert_na){
         #-------------------------------------------------------------------------#